+++ Plugin.cs
@@ -26,6 +26,7 @@ namespace LootManager
         public static HashSet<string> EnabledFilterCategories =
             new HashSet<string>(StringComparer.OrdinalIgnoreCase);
         
+        public static ConfigEntry<KeyboardShortcut> ToggleLootUIHotkey;
         public static ConfigEntry<bool>  AutoLootEnabled;
         public static ConfigEntry<float> AutoLootDistance;
         public static ConfigEntry<string> LootMethod;
@@ -34,6 +35,7 @@ namespace LootManager
         public static ConfigEntry<string> BankLootPageMode;
         public static ConfigEntry<int>   BankPageFirst;
         public static ConfigEntry<int>   BankPageLast;
+        public static ConfigEntry<bool> BankslotAdd;
         public static ConfigEntry<bool>  LootEquipment;
         public static ConfigEntry<EquipmentTierSetting> LootEquipmentTier;
 
@@ -41,8 +43,11 @@ namespace LootManager
         {
             Log = Logger;
             
+            ToggleLootUIHotkey = Config.Bind("Hotkeys", "Toggle Loot UI",new KeyboardShortcut(KeyCode.F6), "Hotkey to toggle the Loot Manager UI window.");
+            
             AutoLootEnabled   = Config.Bind("Autoloot Settings", "Enable Autoloot", true, "Enable or disable auto looting.");
             AutoLootDistance  = Config.Bind("Autoloot Settings", "Autoloot Distance", 20f, "Maximum distance for auto looting.");
+            
             LootMethod        = Config.Bind("Loot Method Settings", "Loot Method", "Blacklist", "Loot method to use: Blacklist, Whitelist, or Standard.");
 
             BankLootEnabled   = Config.Bind("Bankloot Settings", "Bankloot Enabled", false, "If enabled, looted items will be deposited to bank instead of inventory.");
@@ -50,6 +55,7 @@ namespace LootManager
             BankLootPageMode  = Config.Bind("Bankloot Settings", "Bankloot Page Mode", "First Empty", "Mode for depositing items to bank: First Empty or Page Range");
             BankPageFirst     = Config.Bind("Bankloot Settings", "Bank Page First", 20, new ConfigDescription("First bank page to use when in Page Range mode.", new AcceptableValueRange<int>(1, 98)));
             BankPageLast      = Config.Bind("Bankloot Settings", "Bank Page Last", 20, new ConfigDescription("Last bank page to use when in Page Range mode.", new AcceptableValueRange<int>(1, 98)));
+            BankslotAdd       = Config.Bind("Bankloot Settings", "Bankslot Add", false, "If true, items sent to the bank with the inventory Bankslot will be added to the Banklist.");
 
             LootEquipment     = Config.Bind("Filter Settings", "Loot Equipment", true, "If true, loot all equipment.");
             LootEquipmentTier = Config.Bind("Filter Settings", "Loot Equipment Tier", EquipmentTierSetting.All, "Which tiers of equipment to loot: All, Normal Only, Blessed Only, Godly Only, Blessed and Up.");